procmanager_hash.c
This program uses a hashtable data structure to save the commands and make an .out and .err file for each command (based on PID) which records the time info and exit status. If the time is longer than 2 seconds, it will restart the command at the end. 
How to run in terminal: 
./procmanager_hash
sleep 5
ls -latr
sleep 1
^D <---Ctrl-D is the EOF signal that ends the input stream